pyproject.toml
[tool.poetry]
name = "django-issue-tracker"
version = "0.1.0"
description = "Simple issue tracking using django"
authors = ["chen <23054115+cwkang1998@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
python-dotenv = "^0.19.0"
django-import-export = "^2.6.0"
django-bulma = "^0.8.3"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.9b0"
flake8 = "^3.9.2"
selenium = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
